Trait zmq_tokio::SocketRecv

pub trait SocketRecv {
    fn get_rcvmore(&self) -> Result<bool>;

    fn recv(&self, _: &mut Message, _: i32) -> Result<()>;

    fn recv_into(&self, _: &mut [u8], _: i32) -> Result<usize>;

    fn recv_msg(&self, _: i32) -> Result<Message>;

    fn recv_bytes(&self, _: i32) -> Result<Vec<u8>>;

    fn recv_string(&self, _: i32) -> Result<Result<String, Vec<u8>>>;

    fn recv_multipart(&self, _: i32) -> Result<Vec<Vec<u8>>>;
}

API methods for receiving messages with sockets.

Required Methods

fn get_rcvmore(&self) -> Result<bool>

Return true if there are more frames of a multipart message to receive.

fn recv(&self, _: &mut Message, _: i32) -> Result<()>

Receive a message into a Message. The length passed to zmq_msg_recv is the length of the buffer.

fn recv_into(&self, _: &mut [u8], _: i32) -> Result<usize>

Receive bytes into a slice. The length passed to zmq_recv is the length of the slice. The return value is the number of bytes in the message, which may be larger than the length of the slice, indicating truncation.

fn recv_msg(&self, _: i32) -> Result<Message>

Receive a message into a fresh Message.

fn recv_bytes(&self, _: i32) -> Result<Vec<u8>>

Receive a message as a byte vector.

fn recv_string(&self, _: i32) -> Result<Result<String, Vec<u8>>>

Receive a String from the socket.

If the received message is not valid UTF-8, it is returned as the original Vec in the Err part of the inner result.

fn recv_multipart(&self, _: i32) -> Result<Vec<Vec<u8>>>

Receive a multipart message from the socket.

Note that this will allocate a new vector for each message part; for many applications it will be possible to process the different parts sequentially and reuse allocations that way.